requirements.txt
torch~=1.8.1
tqdm~=4.41.1
gdown~=3.13.0
torchvision~=0.9.1
PyYAML~=5.3
Pillow~=6.2.2